Add _misc/.nosearch to exclude directory from feature search
This instructs `normal-top-level-add-subdirs-to-load-path' and other
tools that search a tree for libraries (such as those used by the
Emacsmirror), to ignore that directory.

"readme_update.el" isn't a library, it is a maintenance utility, and
as such it should not provide a feature.  Removing the `provide' from
and using `load' instead of `require' to load it would be another way
to address the issue here addressed with the ".nosearch" file.